fix: O segundo argumento deve ser context: { params: { id: string } }. Usar NextRequest como tipo de request é adequado para funções de rota. context.params.id agora é usado diretamente na função.
