+++ src/app/api/transactions/[id]/route.ts
@@ -1,33 +1,15 @@
-import { NextResponse } from 'next/server'
+import { NextRequest, NextResponse } from 'next/server'
 import { getMongoClient } from '@/db/connectionDb'
+import { getUserIdFromToken } from '@/app/functions/getUserId'
 import { ObjectId } from 'mongodb'
-import jwt from 'jsonwebtoken'
-import { cookies } from 'next/headers'
 
-const JWT_SECRET = process.env.JWT_SECRET || 'your-secret-key'
-
-async function getUserIdFromToken() {
-  const token = (await cookies()).get('auth_token')?.value
-  if (!token) {
-    throw new Error('No token provided')
-  }
-
-  try {
-    const decoded = jwt.verify(token, JWT_SECRET) as { userId: string }
-    return new ObjectId(decoded.userId)
-  } catch (error) {
-    console.error('Invalid token:', error)
-    throw new Error('Invalid token')
-  }
-}
-
-export async function PUT(request: Request, { params }: { params: { id: string } }) {
+export async function PUT(request: NextRequest, context: { params: { id: string } }) {
   try {
     const userId = await getUserIdFromToken()
     const client = await getMongoClient()
     const db = client.db("financeApp")
 
-    const transactionId = new ObjectId(params.id)
+    const transactionId = new ObjectId(context.params.id)
     const updatedTransaction = await request.json()
 
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -55,13 +37,13 @@ export async function PUT(request: Request, { params }: { params: { id: string }
   }
 }
 
-export async function DELETE(request: Request, { params }: { params: { id: string } }) {
+export async function DELETE(request: NextRequest, context: { params: { id: string } }) {
   try {
     const userId = await getUserIdFromToken()
     const client = await getMongoClient()
     const db = client.db("financeApp")
 
-    const transactionId = new ObjectId(params.id)
+    const transactionId = new ObjectId(context.params.id)
 
     const result = await db.collection('transactions').deleteOne({ _id: transactionId, userId })
