+++ src/app/api/transactions/route.ts
@@ -1,25 +1,6 @@
 import { NextResponse } from 'next/server'
 import { getMongoClient } from '@/db/connectionDb'
-import { ObjectId } from 'mongodb'
-import jwt from 'jsonwebtoken'
-import { cookies } from 'next/headers'
-
-const JWT_SECRET = process.env.JWT_SECRET || 'your-secret-key'
-
-async function getUserIdFromToken() {
-  const token = (await cookies()).get('auth_token')?.value
-  if (!token) {
-    throw new Error('No token provided')
-  }
-
-  try {
-    const decoded = jwt.verify(token, JWT_SECRET) as { userId: string }
-    return new ObjectId(decoded.userId)
-  } catch (error) {
-    console.error('Invalid token:', error)
-    throw new Error('Invalid token')
-  }
-}
+import { getUserIdFromToken } from '@/app/functions/getUserId';
 
 export async function GET() {
   try {
